% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_conformal.R
\name{loo_conformal}
\alias{loo_conformal}
\alias{loo_conformal.default}
\alias{loo_conformal.stanreg}
\alias{loo_conformal.brmsfit}
\alias{loo_conformal.array}
\title{Enable jackknife+ predictive intervals for a fit model}
\usage{
loo_conformal(fit, ...)

\method{loo_conformal}{default}(fit, truth, chain = NULL, est_fun = c("mean", "median"), ...)

\method{loo_conformal}{stanreg}(fit, est_fun = c("mean", "median"), ...)

\method{loo_conformal}{brmsfit}(fit, est_fun = c("mean", "median"), ...)

\method{loo_conformal}{array}(
  fit,
  log_lik,
  truth,
  chain = NULL,
  est_fun = c("mean", "median"),
  ...
)
}
\arguments{
\item{fit}{Model fit; an object with \code{posterior_predict()} and \code{log_lik()}
methods. Can also be an \code{array} of posterior predictions.}

\item{...}{Ignored.}

\item{truth}{True values to predict. Not required for \code{rstanarm} or \code{brms}
models.}

\item{chain}{An integer vector identifying the chain numbers for the
posterior draws. Should be provided if multiple chains are used.}

\item{est_fun}{Whether to use the posterior \code{mean} (the default) or \code{median}
as a point estimate.}

\item{log_lik}{If \code{fit} is an array, then \code{log_lik} should be an array of
matching dimension giving the pointwise log-likelihood.}
}
\value{
A modified \code{fit} object with an additional class \code{conformal}.
Calling \code{\link[=predictive_interval]{predictive_interval()}} on this new object will yield conformal
intervals.
}
\description{
Prepares for jackknife+ conformal prediction by performing Pareto-smoothed
importance sampling to yield leave-one-out residuals.
}
\examples{

}
\references{
Vehtari, A., Simpson, D., Gelman, A., Yao, Y., & Gabry, J. (2015).
Pareto smoothed importance sampling. \href{https://arxiv.org/abs/1507.02646}{arXiv preprint arXiv:1507.02646}.
}
