% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal.R
\name{loo_conformal}
\alias{loo_conformal}
\alias{loo_conformal.default}
\alias{loo_conformal.stanreg}
\alias{loo_conformal.brmsfit}
\alias{loo_conformal.array}
\title{Estimate leave-one-out conformal interval widths for a fit model}
\usage{
loo_conformal(fit, ...)

\method{loo_conformal}{default}(fit, truth, est_fun = c("mean", "median"), local = TRUE, ...)

\method{loo_conformal}{stanreg}(fit, est_fun = c("mean", "median"), local = TRUE, ...)

\method{loo_conformal}{brmsfit}(fit, est_fun = c("mean", "median"), local = TRUE, ...)

\method{loo_conformal}{array}(fit, truth, est_fun = c("mean", "median"), local = TRUE, ...)
}
\arguments{
\item{fit}{Model fit}

\item{truth}{True values to predict}
}
\value{

}
\description{
Construct finite-sample calibrated predictive intervals for Bayesian models,
following the approach in Barber et al. (2021). These intervals are
calculated efficiently using importance sampling for the leave-one-out
residuals.  By default, the intervals will also reflect the relative
uncertainty in the Bayesian model, using the locally-weighted conformal
methods of Lei et al. (2018).
}
\examples{

}
\references{
Barber, R. F., Candes, E. J., Ramdas, A., & Tibshirani, R. J. (2021).
Predictive inference with the jackknife+. \emph{The Annals of Statistics, 49}(1),
486-507.

Lei, J., G’Sell, M., Rinaldo, A., Tibshirani, R. J., & Wasserman, L. (2018).
Distribution-free predictive inference for regression. \emph{Journal of the
American Statistical Association, 113}(523), 1094-1111.
}
